Switch to downloading an archive of OpenCV
Instead of using git to checkout a version, we download the artifact &
install it.

This allows us to simplify some of our paths, and make some assumptions.

It also makes it safer to upgrade, as rolling back is a matter of
adjusting the symlink that the ark[opencv] resource creates.